revert back the junit 4 jobTestHelper. it is in another pull request
